requirements.txt
pyyaml
python-twitter
